requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
selenium>=4.0.0
dnspython>=2.0.0 